move test helper functions to adhocracy4 package
