+++ tests/comments/test_templatetags.py
@@ -2,11 +2,11 @@ import json
 import pytest
 import re
 
-from adhocracy4.comments.serializers import ThreadSerializer
 from django.contrib.auth.models import AnonymousUser
 from django.contrib.contenttypes.models import ContentType
 
-from tests import helpers
+from adhocracy4.comments.serializers import ThreadSerializer
+from adhocracy4.test import helpers
 
 
 def react_comment_render_for_props(rf, user, question):
