+++ tests/projects/test_project_views.py
@@ -1,6 +1,7 @@
 import pytest
 from django.core.urlresolvers import reverse
-from tests.helpers import redirect_target
+
+from adhocracy4.test.helpers import redirect_target
 
 
 @pytest.mark.django_db
